requirements.txt
Adafruit_DHT==1.4.0
paho_mqtt==2.1.0
psutil==7.0.0
python-dotenv==1.0.1
smbus2==0.5.0
